refactor type of bhk_type and area

+++ src/Core/Dhanman.MyHome.Application.Contracts/Units/UnitDetailResponse.cs
@@ -12,13 +12,13 @@ public sealed class UnitDetailResponse
     public string UnitType { get; } // Flat Type
     public int OccupantTypeId { get; }
     public string OccupantType { get; }  
-    public float Area { get; }
-    public float BHKType { get; }
+    public decimal Area { get; }
+    public decimal BHKType { get; }
     public Guid AccountId { get; }
     #endregion
 
     #region Constructor
-    public UnitDetailResponse(int id, string name, Guid customerId, int floorId, string floorNumber, int unitTypeId, string unitType, int occupantTypeId, string occupantType, float area, float bHKType, Guid accountId)
+    public UnitDetailResponse(int id, string name, Guid customerId, int floorId, string floorNumber, int unitTypeId, string unitType, int occupantTypeId, string occupantType, decimal area, decimal bHKType, Guid accountId)
     {
         Id = id;
         Name = name;
